refactor profile page
