+++ src/features/profile/index.ts
@@ -1,0 +1,4 @@
+import MyProfile from './Profile'
+import MemberProfile from './memberProfile'
+
+export { MyProfile, MemberProfile }
