--- src/features/profile/myProfile/index.ts
@@ -1,3 +1,0 @@
-import MyProfile from './MyProfile'
-
-export default MyProfile
